+++ src/components/commons/StudyGroup.jsx
